requirements.txt
Django==2.1.7
django-dynamic-fixture==2.0.0
django-rest-framework==0.1.0
djangorestframework==3.9.2
factory-boy==2.12.0
Faker==4.0.1
gunicorn==19.9.0
psycopg2-binary==2.7.7
python-dateutil==2.8.1
pytz==2018.9
six==1.12.0
text-unidecode==1.3
